Extract accessProject
